requirements.txt
fastapi
uvicorn
python-dotenv
requests
pymongo
jinja2
python-telegram-bot